Update SSH action and add MAIL_PASSWORD in Env Var for CI/CD
Updated SSH action to the latest version v1.0.0 and moved environment variables to an earlier stage in the GitHub workflow for the ability to capture mail password from the GitHub secrets for the Continuous Integration / Continuous Deployment (CI/CD) pipeline. This provides better security by not hardcoding the mail password into the workflow.